Moviendo toda la lógica de la base de datos de cotm a un nuevo directorio
